move Pig icon
